山上的风 地心的力 生命向上长成了自己 那时你会看到春野满地
